Fixed issue in server startup due to invalid config provider
