+++ components/auth/org.wso2.carbon.auth.oauth.rest.api/pom.xml
@@ -233,7 +233,7 @@
             org.wso2.carbon.auth.oauth.rest.api.*; version="${carbon.auth.version}"
         </export.package>
         <carbon.component>
-            osgi.service; objectClass="org.wso2.msf4j.Microservice"; serviceCount="2"
+            osgi.service; objectClass="org.wso2.msf4j.Microservice"; serviceCount="3"
         </carbon.component>
     </properties>
 </project>
