+++ components/auth/org.wso2.carbon.auth.oauth.rest.api/src/gen/java/org/wso2/carbon/auth/oauth/rest/api/UserinfoApiService.java
@@ -8,7 +8,6 @@ import org.wso2.msf4j.formparam.FileInfo;
 import org.wso2.msf4j.Request;
 
 import org.wso2.carbon.auth.oauth.rest.api.dto.ErrorDTO;
-import org.wso2.carbon.auth.oauth.rest.api.dto.UserInfoResponseDTO;
 
 import java.util.List;
 import org.wso2.carbon.auth.oauth.rest.api.NotFoundException;
