+++ components/auth/org.wso2.carbon.auth.oauth.rest.api/src/main/java/org/wso2/carbon/auth/oauth/rest/api/impl/UserinfoApiServiceImpl.java
@@ -4,7 +4,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.carbon.auth.oauth.rest.api.NotFoundException;
 import org.wso2.carbon.auth.oauth.rest.api.UserinfoApiService;
-import org.wso2.carbon.auth.oauth.rest.api.dto.UserInfoResponseDTO;
 import org.wso2.carbon.auth.rest.api.commons.dto.ErrorDTO;
 import org.wso2.carbon.auth.rest.api.commons.util.RestApiUtil;
 import org.wso2.carbon.auth.user.info.UserinfoRequestHandler;
@@ -28,11 +27,9 @@ public class UserinfoApiServiceImpl extends UserinfoApiService {
     @Override
     public Response userinfoGet(String authorization, String schema, Request request) throws NotFoundException {
 
-        UserInfoResponseDTO userinfoResponseDTO = new UserInfoResponseDTO();
+        String userInfo;
         try {
-            String userInfo = userinfoRequestHandler.retrieveUserInfo(authorization, schema);
-            userinfoResponseDTO.setUserInfo(userInfo);
-
+            userInfo = userinfoRequestHandler.retrieveUserInfo(authorization, schema);
         } catch (UserInfoException e) {
             String errorMessage = "Error while retrieving user information";
             ErrorDTO errorDTO = RestApiUtil.getErrorDTO(e.getErrorHandler());
@@ -40,7 +37,7 @@ public class UserinfoApiServiceImpl extends UserinfoApiService {
             return Response.status(e.getErrorHandler().getHttpStatusCode()).entity(errorDTO).build();
         }
 
-        return Response.ok().entity(userinfoResponseDTO).build();
+        return Response.ok().entity(userInfo).build();
     }
