+++ components/auth/org.wso2.carbon.auth.user.info/pom.xml
@@ -25,7 +25,7 @@
     <parent>
         <artifactId>auth</artifactId>
         <groupId>org.wso2.carbon.auth</groupId>
-        <version>1.0.39-SNAPSHOT</version>
+        <version>1.0.42-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
